WOW signature verification working

+++ go/types/types.go
@@ -40,8 +40,8 @@ func NewValidatorIndex(value uint16) (ValidatorIndex, error) {
 
 type TicketEntryIndex uint8
 
-func NewTicketEntryIndex(value uint8) (TicketEntryIndex, error) {
-	if value >= uint8(constants.NumTicketEntries) {
+func NewTicketEntryIndex(value uint64) (TicketEntryIndex, error) {
+	if value >= uint64(constants.NumTicketEntries) {
 		return 0, fmt.Errorf("invalid ticket entry index value: must be less than %d", constants.NumTicketEntries)
 	}
 	return TicketEntryIndex(value), nil
